requirements.txt
biopython
click
numpy
matplotlib